+++ packages/plugin-swr/src/generators/queryGenerator.tsx
@@ -93,13 +93,15 @@ export const queryGenerator = createReactGenerator<PluginSwr>({
         )}
         {options.client.importPath ? (
           <>
-            <File.Import name={'fetch'} path={options.client.importPath} />
+            {!shouldUseClientPlugin && <File.Import name={'fetch'} path={options.client.importPath} />}
             <File.Import name={['Client', 'RequestConfig', 'ResponseErrorConfig']} path={options.client.importPath} isTypeOnly />
             {options.client.dataReturnType === 'full' && <File.Import name={['ResponseConfig']} path={options.client.importPath} isTypeOnly />}
           </>
         ) : (
           <>
-            <File.Import name={['fetch']} root={query.file.path} path={path.resolve(config.root, config.output.path, '.kubb/fetch.ts')} />
+            {!shouldUseClientPlugin && (
+              <File.Import name={['fetch']} root={query.file.path} path={path.resolve(config.root, config.output.path, '.kubb/fetch.ts')} />
+            )}
             <File.Import
               name={['Client', 'RequestConfig', 'ResponseErrorConfig']}
               root={query.file.path}